refactor: change components to pages
